+++ app/controllers/entry_pins_controller.rb
@@ -3,5 +3,6 @@ class EntryPinsController < ApplicationController
     @entry = Current.user.entries.find(params[:entry_id])
     Current.user.toggle_entry_pin!(@entry)
     @pinned_count = Current.user.pinned_entry_count
+    @pinned_entries = Current.user.pinned_entries.limit(5)
   end
 end