+++ config/routes.rb
@@ -25,7 +25,7 @@ Rails.application.routes.draw do
     end
   end
 
-  resource :pinned_entry_open, only: [ :create, :destroy ]
+  resource :pinned_entry_open, only: :destroy
 
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html